requirements.txt
Django>=4.2,<5
Pillow
django-extensions
